# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Agentic coding agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Session {
  id: number;
  title: string;
  model: string;
  workspacePath: string;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface SessionInput {
  title: string;
  model?: string;
}

export interface Message {
  id: number;
  sessionId: number;
  role: string;
  content: string;
  /** @nullable */
  toolCalls?: string | null;
  /** @nullable */
  toolCallId?: string | null;
  createdAt: string;
}

export interface SessionWithMessages {
  id: number;
  title: string;
  model: string;
  workspacePath: string;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface ChatInput {
  content: string;
}

export interface FilePathInput {
  path: string;
}

export interface WorkspaceFile {
  name: string;
  path: string;
  type: string;
  size: number;
  modifiedAt: string;
}

export interface WorkspaceFileContent {
  path: string;
  content: string;
  language: string;
}

export interface OllamaModel {
  name: string;
  size: number;
  modifiedAt: string;
}

export interface CredentialsBody {
  /**
     * @minLength 3
     * @maxLength 64
     */
  username: string;
  /**
     * @minLength 8
     * @maxLength 256
     */
  password: string;
}

export interface User {
  id: number;
  username: string;
  createdAt: string;
}

export interface ApiError {
  error: string;
}