# lib/api-zod/src/generated/types/credentialsBody.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Agentic coding agent API
 * OpenAPI spec version: 0.1.0
 */

export interface CredentialsBody {
  /**
     * @minLength 3
     * @maxLength 64
     */
  username: string;
  /**
     * @minLength 8
     * @maxLength 256
     */
  password: string;
}