# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Agentic coding agent API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  createdAt: Date;
}